refactor lotus
