+++ backend/lib/deal-observer.js
@@ -2,7 +2,7 @@
 /** @import {Provider} from 'ethers' */
 
 import { EVENT_TYPES, GLIF_RPC } from './config.js'
-import { ActorEventFilter, LotusService } from './lotus/service.js'
+import { ActorEventFilter, LotusService } from './rpc-service/service.js'
 
 class DealObserver {
   #pgPool
@@ -10,7 +10,7 @@ class DealObserver {
   #rpcProviderURL
   #cache
 
-  constructor (pgPool = null, rpcProviderURL = GLIF_RPC, chainHead = null) {
+  constructor(pgPool = null, rpcProviderURL = GLIF_RPC, chainHead = null) {
     // TODO: Store events in pgPool
     this.#pgPool = pgPool
     this.#rpcProviderURL = rpcProviderURL
@@ -18,7 +18,7 @@ class DealObserver {
     this.#cache.set('chainHead', chainHead)
   }
 
-  async build () {
+  async build() {
     this.#lotusService = await (new LotusService(this.#rpcProviderURL)).build()
     if (!this.#cache.get('chainHead')) {
       const chainHead = await this.#lotusService.getChainHead()
@@ -27,7 +27,7 @@ class DealObserver {
     return this
   }
 
-  async observeBuiltinActorEvents (fromHeight = this.#cache.get('chainHead').Height, toHeight = this.#cache.get('chainHead').Height, eventTypes = EVENT_TYPES) {
+  async observeBuiltinActorEvents(fromHeight = this.#cache.get('chainHead').Height, toHeight = this.#cache.get('chainHead').Height, eventTypes = EVENT_TYPES) {
     return this.#lotusService.getActorEvents(new ActorEventFilter(fromHeight, toHeight, eventTypes))
   }
 }
@@ -37,5 +37,5 @@ export {
 }
 
 (new DealObserver()).build().then((dealObserver) => {
-  dealObserver.observeBuiltinActorEvents(4620803,4620803).then((events) => { console.log(events) })
+  dealObserver.observeBuiltinActorEvents(4620803, 4620803).then((events) => { console.log(events) })
 })
