+++ backend/lib/rpc-service/ipld-schema.js
@@ -3,47 +3,37 @@ import { schemaDmt } from './builtin-actor-events-schemas.js'
 import { base64pad } from 'multiformats/bases/base64'
 import { encode as cborEncode, decode as cborDecode } from '@ipld/dag-cbor'
 
-class Transformer {
+class IpldSchema {
   // A transformer which takes in a json object and returns a typed ClaimEvent object
-  #claimEventTransformer
-  #rawActorEventTransformer
+  #claimEventSchema
+  #rawActorEventSchema
 
-  async build () {
+  async build() {
     // TODO: Catch and log errors
-    this.#claimEventTransformer = this.#createTransformer('ClaimEvent')
-    this.#rawActorEventTransformer = this.#createTransformer('RawActorEvent')
+    this.#claimEventSchema = this.#createType('ClaimEvent')
+    this.#rawActorEventSchema = this.#createType('RawActorEvent')
     return this
   }
 
-  #createTransformer (name) {
+  #createType(name) {
     // TODO: Catch and log errors
     return create(schemaDmt, name)
   }
 
-  transform (typeName, data) {
+  applyType(typeName, data) {
     switch (typeName.toLowerCase()) {
       case 'claimevent':
         // TODO: Catch and log errors
-        return this.#claimEventTransformer.toTyped(data)
+        return this.#claimEventSchema.toTyped(data)
       case 'rawactorevent':
         // TODO: Catch and log errors
-        return this.#rawActorEventTransformer.toTyped(data)
+        return this.#rawActorEventSchema.toTyped(data)
       default:
         console.log(`Unknown type ${typeName}`)
     }
   }
 }
 
-const decodeCborInBase64 = (data) => {
-  return cborDecode(base64pad.baseDecode(data))
-}
-
-const encodeCborInBase64 = (data) => {
-  return base64pad.baseEncode(cborEncode(data))
-}
-
 export {
-  decodeCborInBase64,
-  encodeCborInBase64,
-  Transformer
+  IpldSchema
 }
