+++ backend/test/observer.test.js
@@ -1,11 +1,10 @@
 import { after, before, beforeEach, describe, it, mock } from 'node:test'
 import { createPgPool, migrateWithPgClient } from '@filecoin-station/deal-observer-db'
-import { Transformer } from '../lib/lotus/transform.js'
+import { IpldSchema } from '../lib/rpc-service/ipld-schema.js'
 import assert from 'assert'
 import { claimTestEvent } from './test_data/claimEvent.js'
-import { LotusService } from '../lib/lotus/service.js'
+import { LotusService } from '../lib/rpc-service/service.js'
 import { GLIF_RPC } from '../lib/config.js'
-import { startServer,testServerURL } from './mockServer.js'
 describe('deal-observer-backend', () => {
   let pgPool
 
@@ -37,8 +36,8 @@ describe('deal-observer-backend', () => {
 
   describe('Transformer', () => {
     it('transforms a claim event payload to a typed object', async () => {
-      const transformer = await (new Transformer().build())
-      const transformedClaimEvent = transformer.transform('ClaimEvent', claimTestEvent)
+      const transformer = await (new IpldSchema().build())
+      const transformedClaimEvent = transformer.applyType('ClaimEvent', claimTestEvent)
       assert(transformedClaimEvent !== undefined, 'transformedClaimEvent is undefined')
       assert.deepStrictEqual(transformedClaimEvent, claimTestEvent)
     })
@@ -46,21 +45,8 @@ describe('deal-observer-backend', () => {
 
   describe('LotusService', () => {
     let lotusService
-    let server 
-    before(async () => {
-      server = await startServer()
-      lotusService = await (new LotusService(testServerURL)).build()
-      }) 
-    after(async () => {
-      server.close();
-    })
 
     it('test the retrieval of built in actor events', async () => {
-      console.log(await fetch('http://localhost:8080', {
-        method: 'GET',
-      }));
-      //let response = await lotusService.getChainHead()
-      //console.log(response)
     })
   })
 })
